Move utilities + add some tests
