+++ src/ir/auth_logic/lowering_ast_datalog.h
@@ -28,7 +28,7 @@
 #include "src/common/logging/logging.h"
 #include "src/ir/auth_logic/ast.h"
 #include "src/ir/auth_logic/datalog_ir.h"
-#include "src/ir/auth_logic/move_append.h"
+#include "src/utils/move_append.h"
 #include "src/utils/overloaded.h"
 
 namespace raksha::ir::auth_logic {
